backend/requirements.txt
flask[async]>=2.3.3
flask-cors>=4.0.0
flask-socketio>=5.3.6
python-socketio>=5.8.0
requests>=2.31.0
aiohttp>=3.8.5
anthropic>=0.7.7
openai>=1.33.0
google-generativeai>=0.3.0
mem0ai>=0.0.5
python-dotenv>=1.0.0
redis>=4.5.0
scrapybara>=1.0.0
beautifulsoup4>=4.12.2
websockets>=11.0.2
pydantic>=2.5.0
dataclasses-json>=0.6.0
asyncio-throttle>=1.0.2
psutil>=5.9.0
colorama>=0.4.6
rich>=13.0.0
e2b-code-interpreter>=1.5.1
google-cloud-aiplatform>=1.38.0
google-auth>=2.22.0

# New dependencies for Agentic Superpowers V3.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# New dependencies for Collaborative Workspaces V3.0
python-socketio[client]>=5.8.0
eventlet>=0.33.0

# Enhanced async and concurrency support
asyncio>=3.4.3
